+++ Release Notes.md
@@ -7,7 +7,8 @@ This release wraps more logic within os checks to make it build on tvOS and watc
 
 The release also adds:
 
-* A new `CircularProgressBar` which renders a progress in a circle.
+* A new `CircularProgressBar` which renders a percentual progress in a circle.
+* A new `CircularProgressView` which renders a spinner.
 * A new `FetchedDataView` view that shows a spinner, fetched data or a failure view.
 * A new `FullScreenCover` namespace that lets you apply full screen covers like sheets, alerts and toasts.